@@ -10,60 +10,58 @@
     "lint": "ng lint",
     "e2e": "ng e2e",
     "generate-types": "gql-gen --config codegen.yml",
-    "dev:ssr": "ng run reference-storefront:serve-ssr",
-    "serve:ssr": "node dist/server/main.js",
-    "build:ssr": "ng build --configuration=production && ng run reference-storefront:server:production",
     "build:ssr:docker": "ng build --configuration=production,docker && ng run reference-storefront:server:docker",
-    "prerender": "ng run reference-storefront:prerender"
+    "serve:ssr:reference-storefront": "node dist/server/server.mjs"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.0.4",
-    "@angular/cdk": "^16.0.3",
-    "@angular/common": "^16.0.4",
-    "@angular/compiler": "^16.0.4",
-    "@angular/core": "^16.0.4",
-    "@angular/forms": "^16.0.4",
-    "@angular/platform-browser": "^16.0.4",
-    "@angular/platform-browser-dynamic": "^16.0.4",
-    "@angular/platform-server": "^16.0.4",
-    "@angular/router": "^16.0.4",
-    "@angular/service-worker": "^16.0.4",
-    "@apollo/client": "^3.7.15",
-    "@fortawesome/angular-fontawesome": "^0.13.0",
+    "@angular/animations": "^20.3.9",
+    "@angular/cdk": "^20.2.10",
+    "@angular/common": "^20.3.9",
+    "@angular/compiler": "^20.3.9",
+    "@angular/core": "^20.3.9",
+    "@angular/forms": "^20.3.9",
+    "@angular/platform-browser": "^20.3.9",
+    "@angular/platform-browser-dynamic": "^20.3.9",
+    "@angular/platform-server": "^20.3.9",
+    "@angular/router": "^20.3.9",
+    "@angular/service-worker": "^20.3.9",
+    "@angular/ssr": "^20.3.8",
+    "@apollo/client": "^4.0.8",
+    "@fortawesome/angular-fontawesome": "^2.0.1",
     "@fortawesome/fontawesome-svg-core": "^6.2.1",
     "@fortawesome/free-brands-svg-icons": "^6.2.1",
     "@fortawesome/free-solid-svg-icons": "^6.2.1",
-    "@nguniversal/express-engine": "^16.0.2",
+    "@primeuix/themes": "^1.2.5",
     "@tailwindcss/forms": "^0.5.3",
-    "apollo-angular": "^5.0.0",
+    "apollo-angular": "^12.1.0",
     "compression": "^1.7.4",
     "core-js": "^3.6.5",
-    "express": "^4.15.2",
+    "express": "^5.1.0",
     "graphql": "^16.6.0",
     "photoswipe": "^5.3.4",
-    "rxjs": "~6.6.3",
+    "primeng": "^20.2.0",
+    "rxjs": "^7.4.0",
     "tslib": "^2.4.0",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.0.5",
-    "@angular-eslint/builder": "16.1.0",
-    "@angular-eslint/eslint-plugin": "16.1.0",
-    "@angular-eslint/eslint-plugin-template": "16.1.0",
-    "@angular-eslint/schematics": "16.1.0",
-    "@angular-eslint/template-parser": "16.1.0",
-    "@angular/cli": "^16.0.5",
-    "@angular/compiler-cli": "^16.0.4",
-    "@angular/language-service": "^16.0.4",
+    "@angular-eslint/builder": "^20.5.0",
+    "@angular-eslint/eslint-plugin": "^20.5.0",
+    "@angular-eslint/eslint-plugin-template": "^20.5.0",
+    "@angular-eslint/schematics": "^20.5.0",
+    "@angular-eslint/template-parser": "^20.5.0",
+    "@angular/build": "^20.3.8",
+    "@angular/cli": "^20.3.8",
+    "@angular/compiler-cli": "^20.3.9",
+    "@angular/language-service": "^20.3.9",
     "@graphql-codegen/add": "^3.2.1",
     "@graphql-codegen/cli": "^2.14.0",
     "@graphql-codegen/fragment-matcher": "^3.3.1",
     "@graphql-codegen/typescript": "^2.8.2",
     "@graphql-codegen/typescript-operations": "^2.5.7",
-    "@nguniversal/builders": "^16.0.2",
     "@types/compression": "^1.7.0",
-    "@types/express": "^4.17.8",
+    "@types/express": "^5.0.1",
     "@types/graphql": "^14.0.5",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.6",
@@ -71,20 +69,21 @@
     "@typescript-eslint/eslint-plugin": "5.62.0",
     "@typescript-eslint/parser": "5.62.0",
     "autoprefixer": "^10.4.13",
+    "browser-sync": "^3.0.0",
     "codelyzer": "^6.0.0",
     "eslint": "^8.44.0",
-    "jasmine-core": "~3.6.0",
+    "jasmine-core": "^5.0.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.2",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^16.0.1",
+    "karma": "^6.4.4",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
+    "ng-packagr": "^20.3.0",
     "postcss": "^8.4.18",
     "tailwindcss": "^3.2.1",
     "ts-loader": "^8.0.4",
     "ts-node": "~9.0.0",
-    "typescript": "~5.0.4"
+    "typescript": "^5.9.0"
   }
 }
