@@ -1,6 +1,6 @@
 import { isPlatformBrowser } from '@angular/common';
 import { HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest, HttpResponse } from '@angular/common/http';
-import { Inject, Injectable, Injector, PLATFORM_ID } from '@angular/core';
+import { Injectable, Injector, PLATFORM_ID, inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
@@ -15,12 +15,11 @@ import { DataService } from './data.service';
  */
 @Injectable()
 export class DefaultInterceptor implements HttpInterceptor {
-    constructor(
-        private dataService: DataService,
-        private injector: Injector,
-        private router: Router,
-        @Inject(PLATFORM_ID) private platformId: any,
-    ) {}
+    private dataService = inject(DataService);
+    private injector = inject(Injector);
+    private router = inject(Router);
+    private platformId = inject(PLATFORM_ID);
+
 
     intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
         return next.handle(req).pipe(
