@@ -1,6 +1,6 @@
 import { Overlay, OverlayConfig } from '@angular/cdk/overlay';
-import { ComponentPortal, PortalInjector } from '@angular/cdk/portal';
-import { Injectable, Injector, Type } from '@angular/core';
+import { ComponentPortal } from '@angular/cdk/portal';
+import { Injectable, Injector, Type, inject } from '@angular/core';
 import { Observable, race } from 'rxjs';
 import { finalize, mapTo, take, tap } from 'rxjs/operators';
 
@@ -14,7 +14,9 @@ import { Dialog, DIALOG_COMPONENT, MODAL_OPTIONS, ModalOptions } from './modal-t
  */
 @Injectable({ providedIn: 'root' })
 export class ModalService {
-    constructor(private overlay: Overlay, private injector: Injector) {}
+    private overlay = inject(Overlay);
+    private injector = inject(Injector);
+
 
     /**
      * Create a modal from a component. The component must implement the {@link Dialog} interface.
@@ -79,17 +81,19 @@ export class ModalService {
         });
         const backdropClick$ = overlayRef.backdropClick().pipe(mapTo(undefined));
 
-        return race<R | undefined>(close$, backdropClick$).pipe(
+        return race(close$, backdropClick$).pipe(
             take(1),
             finalize(() => overlayRef.dispose()),
         );
     }
 
-    private createInjector<T, R>(component: Type<T> & Type<Dialog<R>>,  options?: ModalOptions<T>): PortalInjector {
-        const weakMap = new WeakMap<any, any>([
-            [DIALOG_COMPONENT, component],
-            [MODAL_OPTIONS, options],
-        ]);
-        return new PortalInjector(this.injector, weakMap);
+    private createInjector<T, R>(component: Type<T> & Type<Dialog<R>>,  options?: ModalOptions<T>): Injector {
+        return Injector.create({
+            parent: this.injector,
+            providers: [
+                { provide: DIALOG_COMPONENT, useValue: component },
+                { provide: MODAL_OPTIONS, useValue: options },
+            ],
+        });
     }
 }
