@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit, inject } from '@angular/core';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
@@ -7,15 +7,17 @@ import { GET_CUSTOMER_ADDRESSES } from '../../../common/graphql/documents.graphq
 import { DataService } from '../../../core/providers/data/data.service';
 
 @Component({
-    selector: 'vsf-account-address-book',
+  standalone: false,
+  selector: 'vsf-account-address-book',
     templateUrl: './account-address-book.component.html',
     // styleUrls: ['./account-address-book.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AccountAddressBookComponent implements OnInit {
+    private dataService = inject(DataService);
+
 
     addresses$: Observable<NonNullable<GetCustomerAddressesQuery['activeCustomer']>['addresses'] | undefined>;
-    constructor(private dataService: DataService) { }
 
     ngOnInit() {
         this.addresses$ = this.dataService.query<GetCustomerAddressesQuery>(GET_CUSTOMER_ADDRESSES).pipe(
