@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit, inject } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Observable } from 'rxjs';
 import { filter, map, switchMap } from 'rxjs/operators';
@@ -10,15 +10,18 @@ import { DataService } from '../../../core/providers/data/data.service';
 import { GET_ORDER } from './account-order-detail.graphql';
 
 @Component({
-    selector: 'vsf-account-order-detail',
+  standalone: false,
+  selector: 'vsf-account-order-detail',
     templateUrl: './account-order-detail.component.html',
     styleUrls: ['./account-order-detail.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AccountOrderDetailComponent implements OnInit {
+    private dataService = inject(DataService);
+    private route = inject(ActivatedRoute);
+
 
     order$: Observable<GetOrderQuery['orderByCode'] | undefined>;
-    constructor(private dataService: DataService, private route: ActivatedRoute) { }
 
     ngOnInit() {
         this.order$ = this.route.paramMap.pipe(
