@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit, inject } from '@angular/core';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
@@ -8,15 +8,17 @@ import { DataService } from '../../../core/providers/data/data.service';
 import { GET_ORDER_LIST } from './account-order-list.graphql';
 
 @Component({
-    selector: 'vsf-account-order-list',
+  standalone: false,
+  selector: 'vsf-account-order-list',
     templateUrl: './account-order-list.component.html',
     // styleUrls: ['./account-order-list.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AccountOrderListComponent implements OnInit {
+    private dataService = inject(DataService);
+
 
     orders$: Observable<NonNullable<GetOrderListQuery['activeCustomer']>['orders']['items'] | undefined>;
-    constructor(private dataService: DataService) { }
 
     ngOnInit() {
         this.orders$ = this.dataService.query<GetOrderListQuery, GetOrderListQueryVariables>(GET_ORDER_LIST, {
