@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit, inject } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
 import { RequestPasswordResetMutation, RequestPasswordResetMutationVariables } from '../../../common/generated-types';
@@ -7,16 +7,20 @@ import { DataService } from '../../../core/providers/data/data.service';
 import { REQUEST_PASSWORD_RESET } from './forgotten-password.graphql';
 
 @Component({
-    selector: 'vsf-forgotten-password',
+  standalone: false,
+  selector: 'vsf-forgotten-password',
     templateUrl: './forgotten-password.component.html',
     styleUrls: ['./forgotten-password.component.scss'],
     changeDetection: ChangeDetectionStrategy.Default,
 })
 export class ForgottenPasswordComponent {
+    private dataService = inject(DataService);
+    private route = inject(ActivatedRoute);
+
     emailAddress = '';
     submitted = false;
 
-    constructor(private dataService: DataService, private route: ActivatedRoute) {
+    constructor() {
         this.emailAddress = this.route.snapshot.paramMap.get('email') ?? '';
     }
 
