@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit, inject } from '@angular/core';
 
 import { RegisterMutation, RegisterMutationVariables } from '../../../common/generated-types';
 import { DataService } from '../../../core/providers/data/data.service';
@@ -6,18 +6,20 @@ import { DataService } from '../../../core/providers/data/data.service';
 import { REGISTER } from './register.graphql';
 
 @Component({
-    selector: 'vsf-register',
+  standalone: false,
+  selector: 'vsf-register',
     templateUrl: './register.component.html',
     // styleUrls: ['./register.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class RegisterComponent {
+    private dataService = inject(DataService);
+    private changeDetector = inject(ChangeDetectorRef);
+
     firstName: string;
     lastName: string;
     emailAddress: string;
     registrationSent = false;
-    constructor(private dataService: DataService,
-                private changeDetector: ChangeDetectorRef) { }
 
     register() {
         this.dataService.mutate<RegisterMutation, RegisterMutationVariables>(REGISTER, {
