@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
@@ -7,8 +7,8 @@ import { StateService } from '../../core/providers/state/state.service';
 
 @Injectable({ providedIn: 'root' })
 export class SignInGuard  {
+    private stateService = inject(StateService);
 
-    constructor(private stateService: StateService) {}
 
     canActivate(route: ActivatedRouteSnapshot): Observable<boolean> {
         return this.stateService.select(state => state.signedIn).pipe(
