@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { provideClientHydration } from '@angular/platform-browser';
 import { ServerModule } from '@angular/platform-server';
 
 import { AppComponent } from './app.component';
@@ -11,9 +10,7 @@ import { AppModule } from './app.module';
         ServerModule,
     ],
     bootstrap: [AppComponent],
-    providers: [
-        provideClientHydration(),
-    ],
+    providers: [],
 })
 export class AppServerModule {
 }
