@@ -1,7 +1,8 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 
 @Component({
-    selector: 'vsf-checkout-stage-indicator',
+  standalone: false,
+  selector: 'vsf-checkout-stage-indicator',
     templateUrl: './checkout-stage-indicator.component.html',
     // styleUrls: ['./checkout-stage-indicator.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
