@@ -3,7 +3,8 @@ import { ChangeDetectionStrategy, Component, Input, OnInit } from '@angular/core
 import { GetCollectionQuery } from '../../../common/generated-types';
 
 @Component({
-    selector: 'vsf-collection-breadcrumbs',
+  standalone: false,
+  selector: 'vsf-collection-breadcrumbs',
     templateUrl: './collection-breadcrumbs.component.html',
     styleUrls: ['./collection-breadcrumbs.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
