@@ -1,7 +1,8 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
-    selector: 'vsf-layout',
+  standalone: false,
+  selector: 'vsf-layout',
     templateUrl: './layout.component.html',
     // styleUrls: ['./layout.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
