@@ -1,13 +1,15 @@
-import { Directive, OnInit, TemplateRef } from '@angular/core';
+import { Directive, OnInit, TemplateRef, inject } from '@angular/core';
 
 import { ModalDialogComponent } from './modal-dialog.component';
 
 /**
  * A helper directive used to correctly embed the modal buttons in the {@link ModalDialogComponent}.
  */
-@Directive({ selector: '[vsfDialogButtons]' })
+@Directive({  standalone: false, selector: '[vsfDialogButtons]' })
 export class DialogButtonsDirective implements OnInit {
-    constructor(private modal: ModalDialogComponent<any>, private templateRef: TemplateRef<any>) {}
+    private modal = inject<ModalDialogComponent<any>>(ModalDialogComponent);
+    private templateRef = inject<TemplateRef<any>>(TemplateRef);
+
 
     ngOnInit() {
         this.modal.registerButtonsTemplate(this.templateRef);
