@@ -1,29 +1,19 @@
-import {
-    Component,
-    ComponentFactoryResolver,
-    EventEmitter,
-    Input,
-    OnInit,
-    Output,
-    Type,
-    ViewContainerRef,
-} from '@angular/core';
+import { Component, ComponentFactoryResolver, EventEmitter, Input, OnInit, Output, Type, ViewContainerRef, inject } from '@angular/core';
 
 /**
  * A helper component used to embed a component instance into the {@link ModalDialogComponent}
  */
 @Component({
-    selector: 'vsf-dialog-component-outlet',
+  standalone: false,
+  selector: 'vsf-dialog-component-outlet',
     template: ``,
 })
 export class DialogComponentOutletComponent implements OnInit {
+    private viewContainerRef = inject(ViewContainerRef);
+    private componentFactoryResolver = inject(ComponentFactoryResolver);
+
     @Input() component: Type<any>;
     @Output() create = new EventEmitter<any>();
-
-    constructor(
-        private viewContainerRef: ViewContainerRef,
-        private componentFactoryResolver: ComponentFactoryResolver,
-    ) {}
 
     ngOnInit() {
         const factory = this.componentFactoryResolver.resolveComponentFactory(this.component);
