@@ -2,6 +2,7 @@ import { Component, OnInit, ChangeDetectionStrategy, Input } from '@angular/core
 import { CartFragment } from '../../../common/generated-types';
 
 @Component({
+  standalone: false,
   selector: 'vsf-cart-totals',
   templateUrl: './cart-totals.component.html',
   styleUrls: ['./cart-totals.component.scss'],
@@ -9,7 +10,7 @@ import { CartFragment } from '../../../common/generated-types';
 })
 export class CartTotalsComponent implements OnInit {
     @Input() cart: CartFragment;
-  constructor() { }
+  constructor() {}
 
   ngOnInit(): void {
   }
