@@ -1,7 +1,8 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 
 @Component({
-    selector: 'vsf-centered-card',
+  standalone: false,
+  selector: 'vsf-centered-card',
     templateUrl: './centered-card.component.html',
     styleUrls: ['./centered-card.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
