@@ -1,20 +1,21 @@
-import { ChangeDetectionStrategy, Component, Input, OnChanges } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Input, OnChanges, inject } from '@angular/core';
 import { DomSanitizer, SafeStyle } from '@angular/platform-browser';
 
 import { GetCollectionQuery } from '../../../common/generated-types';
 import { AssetPreviewPipe } from '../../pipes/asset-preview.pipe';
 
 @Component({
-    selector: 'vsf-collection-card',
+  standalone: false,
+  selector: 'vsf-collection-card',
     templateUrl: './collection-card.component.html',
     // styleUrls: ['./collection-card.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CollectionCardComponent implements OnChanges {
+    private sanitizer = inject(DomSanitizer);
+
     @Input() collection: NonNullable<GetCollectionQuery['collection']>;
     backgroundImage: SafeStyle;
-
-    constructor(private sanitizer: DomSanitizer) {}
 
     ngOnChanges() {
         if (this.collection.featuredAsset) {
