@@ -1,26 +1,16 @@
-import {
-    ChangeDetectionStrategy,
-    ChangeDetectorRef,
-    Component,
-    ContentChild,
-    EventEmitter,
-    Input,
-    OnChanges,
-    OnDestroy,
-    OnInit,
-    Output,
-    SimpleChanges,
-    TemplateRef,
-} from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, ContentChild, EventEmitter, Input, OnChanges, OnDestroy, OnInit, Output, SimpleChanges, TemplateRef, inject } from '@angular/core';
 import { Subject, Subscription } from 'rxjs';
 import { debounceTime, throttleTime } from 'rxjs/operators';
 
 @Component({
-    selector: 'vsf-radio-card-fieldset',
+  standalone: false,
+  selector: 'vsf-radio-card-fieldset',
     template: `<fieldset><ng-content></ng-content></fieldset> `,
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class RadioCardFieldsetComponent<T = any> implements OnInit, OnChanges, OnDestroy {
+    private changeDetector = inject(ChangeDetectorRef);
+
     @Input() selectedItemId: string;
     @Input() idFn: (item: T) => string;
     @Output() selectItem = new EventEmitter<T>();
@@ -28,8 +18,6 @@ export class RadioCardFieldsetComponent<T = any> implements OnInit, OnChanges, O
     focussedId: string | undefined = undefined;
     private idChange$ = new Subject<T>();
     private subscription: Subscription;
-
-    constructor(private changeDetector: ChangeDetectorRef) {}
 
     ngOnInit() {
         this.subscription = this.idChange$
